requirements.txt
django-anymail==8.6
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.12.0
django-dotenv==1.4.2
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
django-templated-email==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docutils==0.16
drf-yasg==1.20.0
gunicorn==20.1.0
pillow==9.3.0
whitenoise==6.2.0